+++ src/onemail/om-lambda-config-set-processor/CHANGELOG.md
@@ -1,5 +1,17 @@
 # om-lambda-config-set-processor
 
+## 1.5.0
+
+### Minor Changes
+
+- defc523: Skip update if current and new state are equal to QUEUED
+
+## 1.4.3
+
+### Patch Changes
+
+- bcd64ba: Reject event handling
+
 ## 1.4.2
 
 ### Patch Changes